Move entity manager out of main class
